# Conflicts:
#	src/main/java/com/example/springboot/repositories/UserRepository.java
#	src/main/java/com/example/springboot/services/AccountService.java
#	src/main/java/com/example/springboot/services/QuestionService.java
#	target/classes/com/example/springboot/ZappyApplication.class
#	target/classes/com/example/springboot/controllers/AchievementController.class
#	target/classes/com/example/springboot/controllers/AlphabetController.class
#	target/classes/com/example/springboot/controllers/AnswerController.class
#	target/classes/com/example/springboot/controllers/GrammarController.class
#	target/classes/com/example/springboot/controllers/KanjiController.class
#	target/classes/com/example/springboot/controllers/LessonController.class
#	target/classes/com/example/springboot/controllers/VocabularyController.class
#	target/classes/com/example/springboot/converters/AlphabetConverter.class
#	target/classes/com/example/springboot/converters/GrammarConverter.class
#	target/classes/com/example/springboot/converters/KanjiConverter.class
#	target/classes/com/example/springboot/converters/LessonConverter.class
#	target/classes/com/example/springboot/converters/VocabularyConverter.class
#	target/classes/com/example/springboot/entities/AccountEntity.class
#	target/classes/com/example/springboot/entities/AchievenmentEntity.class
#	target/classes/com/example/springboot/entities/AlphabetEntity.class
#	target/classes/com/example/springboot/entities/AnswerEntity.class
#	target/classes/com/example/springboot/entities/GameEntity.class
#	target/classes/com/example/springboot/entities/GrammarEntity.class
#	target/classes/com/example/springboot/entities/HistoryEntity.class
#	target/classes/com/example/springboot/entities/KanjiEntity.class
#	target/classes/com/example/springboot/entities/LessonEntity.class
#	target/classes/com/example/springboot/entities/MatchWordEntity.class
#	target/classes/com/example/springboot/entities/QuestionTypeEntity.class
#	target/classes/com/example/springboot/entities/RoleEntity.class
#	target/classes/com/example/springboot/entities/RoomEntity.class
#	target/classes/com/example/springboot/entities/SkillEntity.class
#	target/classes/com/example/springboot/entities/VocabularyEntity.class
#	target/classes/com/example/springboot/exception/ResourceNotFoundException.class
#	target/classes/com/example/springboot/repositories/AchievenmentRepository.class
#	target/classes/com/example/springboot/repositories/AlphabetRepository.class
#	target/classes/com/example/springboot/repositories/AnswerRepository.class
#	target/classes/com/example/springboot/repositories/GameRepository.class
#	target/classes/com/example/springboot/repositories/GrammarRepository.class
#	target/classes/com/example/springboot/repositories/HistoryRepository.class
#	target/classes/com/example/springboot/repositories/KanjiRepository.class
#	target/classes/com/example/springboot/repositories/LessonRepository.class
#	target/classes/com/example/springboot/repositories/MatchWordRepository.class
#	target/classes/com/example/springboot/repositories/QuestionTypeRepository.class
#	target/classes/com/example/springboot/repositories/RoleRepository.class
#	target/classes/com/example/springboot/repositories/RoomRepository.class
#	target/classes/com/example/springboot/repositories/SkillRepository.class
#	target/classes/com/example/springboot/repositories/UserRepository.class
#	target/classes/com/example/springboot/repositories/VocabularyRepository.class
#	target/classes/com/example/springboot/services/AchievenmentService.class
#	target/classes/com/example/springboot/services/AlphabetService.class
#	target/classes/com/example/springboot/services/AnswerService.class
#	target/classes/com/example/springboot/services/GrammarService.class
#	target/classes/com/example/springboot/services/HistoryService.class
#	target/classes/com/example/springboot/services/KanjiService.class
#	target/classes/com/example/springboot/services/LessonService.class
#	target/classes/com/example/springboot/services/MatchWordService.class
#	target/classes/com/example/springboot/services/QuestionService.class
#	target/classes/com/example/springboot/services/QuestionTypeService.class
#	target/classes/com/example/springboot/services/RoleService.class
#	target/classes/com/example/springboot/services/RoomService.class
#	target/classes/com/example/springboot/services/SkillService.class
#	target/classes/com/example/springboot/services/UserService.class
#	target/classes/com/example/springboot/services/VocabularyService.class

+++ src/main/java/com/example/springboot/entities/GameRecordEntity.java
@@ -29,7 +29,7 @@ public class GameRecordEntity {
 	private Long timePlayed;
 	
 	@Column(name = "score")
-	private Double score;
+	private Long score;
 	
 	@Column(name = "game_name")
 	private String gameName;
@@ -41,7 +41,7 @@ public class GameRecordEntity {
 		this.timeCreated = Date.valueOf(LocalDate.now(ZoneId.of("GMT+07:00")));
 	}
 
-	public GameRecordEntity(Long id, Long timePlayed, Double score, String gameName, Long userId) {
+	public GameRecordEntity(Long id, Long timePlayed, Long score, String gameName, Long userId) {
 		super();
 		this.id = id;
 		this.timeCreated = Date.valueOf(LocalDate.now(ZoneId.of("GMT+07:00")));;
@@ -75,11 +75,11 @@ public class GameRecordEntity {
 		this.timePlayed = timePlayed;
 	}
 
-	public Double getScore() {
+	public Long getScore() {
 		return score;
 	}
 
-	public void setScore(Double score) {
+	public void setScore(Long score) {
 		this.score = score;
 	}
